storage: Separate into multiple crates
Now we have:
* `storage-core` with the backend-agnostic abstractions
* `storage-inmemory` with the in-memory implementation
* `storage` that collects and re-exports the above two

When using the top-level `storage`, the interface remains unchanged.
