+++ storage/core/src/traits.rs
