+++ storage/inmemory/Cargo.toml
@@ -1,0 +1,11 @@
+[package]
+name = "storage-inmemory"
+version = "0.1.0"
+edition = "2021"
+license = "MIT"
+
+[dependencies]
+
+# Local dependencies
+common = { path = "../../common" }
+storage-core = { path = "../core" }
